add encoding

+++ runbook.py
@@ -7,5 +7,6 @@ import shutil
 if __name__ == "__main__":
     this_dir = Path().resolve()
     build_dir = this_dir / "testbook/_build"
-    shutil.rmtree(build_dir)
+    if build_dir.is_dir():
+        shutil.rmtree(build_dir)
     main()
